requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.2
flask-admin==1.6.1
pyvis==0.3.2
networkx==3.2.1
python-dotenv==1.0.0 